+++ crates/core/tests/java/block_comment_at_first_column/indented_true.out.java
@@ -1,0 +1,7 @@
+class A {
+    void m() {
+        int x = 1;
+/* an indented block comment */
+        int y = 2;
+    }
+}