+++ crates/core/tests/java/line_comment_at_first_column/indented.java
@@ -1,0 +1,7 @@
+class A {
+    void m() {
+        int x = 1;
+        // an indented line comment
+        int y = 2;
+    }
+}