+++ crates/core/tests/java/wrap_comments/long_block_wrapped.out.java
@@ -1,0 +1,8 @@
+class A {
+    void m() {
+        /* alpha beta gamma delta
+         * epsilon zeta eta theta iota
+         * kappa lambda mu */
+        int x = 1;
+    }
+}